# Repositories
Repositories are how we organize our code. They contain all the functionality of our code. Repositories have two great qualities: collaboration and history.

## Collaboration
A repository will usually have a singular place to store files in the cloud. Repositories tend to also come with other tools that double down in collaboration. We will see that the codebase you are currently navigating through

## History
Repositories have history. They change over time. In fact, they are the composition of changes over time. Each of these changes is called a commit. A commit often comes with a unique ID that represent the commit. When we look at a commit, we often see the composition of all commits before it.

The lesson structure in this program is aligned with how repositories change over time. Each lesson represents a commit. Overtime this will help with the understanding of what a repository is.

## Further Reading
 - https://en.wikipedia.org/wiki/Software_repository