# READMEs
READMEs are documents that explain a directories contents. When looking through a new directory, always go to the README and read its content. It will give you a highlevel understanding of the directories contents. They are usually written as markdown files.

In Github, they get displayed when viewing the directory.

## Intro documents
Intro documents like README are the go-to documents when entering a repository. They are usually how a program will find the content of the directory.

In javascript, they are named `index.js`
In Python, they are name `__init__.py`

## Setup and Verify
READMEs serve a great purpose of helping a developer know how to setup a document. It's a good pattern to include a verification step for any setup step. This will allow developers to know that they have succeeded in implementing a step, and often, to skip a step that they have already implemented for another context.