The first thing to realize is that a website is just a document; however, it is not just any document. Like all software documents, it exists in a context. This context is a web browser. A web browser handles two important processes. First, the web browser retrieves a document, often from another computer. Second it renders that document according to an agreed upon specification.
The specification for how to render an HTML document is actually not prescribed by a single entity, but rather a group of different entities that reach a consensus. These entities are the different companies that make web browsers (Google makes Chrome, Apple makes Safari, etc.) The Mozilla Foundation serves to document the consensus. That documentation can be found here: https://developer.mozilla.org/en-US/docs/Web/HTML . Not all browsers support the same set of features. For a long time, web developers hated working with Internet Explorer, because it did not support newer, well-liked features.

The standard name for an HTML document is index.html. When programs look for web documents, they will often look for this document by default.

The index.html in this lesson is the simplest form of an html document. There are a few key features to notice.